issue-18 Synchronizer を partial classs 化
